requirements.txt
Flask==1.1.2
pylint==2.5.3
pytest==5.3.2; python_version > "3.0"
pytest==5.3.2; python_version < "3.0"
gunicorn==20.0.4
google-cloud-secret-manager==1.0.0
python-dotenv==0.13.0
requests==2.23.0
requests-mock==1.8.0
google-cloud-monitoring==1.0.0
